refactor: move features/dashboard -> entrypoints/dashboard
