+++ src/entrypoints/dashboard/client/components/activity.jsx
