+++ src/entrypoints/dashboard/client/components/header.jsx
