+++ src/entrypoints/dashboard/client/components/modal.jsx
