+++ src/entrypoints/dashboard/client/components/toolbar.jsx
