+++ src/entrypoints/dashboard/client/components/worktree-card.jsx
