+++ src/entrypoints/dashboard/client/components/worktree-sections.jsx
