+++ src/entrypoints/dashboard/client/index.html
