+++ src/entrypoints/dashboard/client/lib/dashboard-session.js
