+++ src/entrypoints/dashboard/client/lib/preferences.js
