+++ src/entrypoints/dashboard/client/lib/worktree-presentation.js
