+++ src/entrypoints/dashboard/dashboard-actions.json
