+++ src/entrypoints/dashboard/dashboard-mcp-routes.ts
@@ -1,7 +1,7 @@
 import type http from 'node:http';
 
-import {formatMcpDoctor, runMcpDoctor} from '../mcp-server/mcp-server-doctor.js';
-import {formatMcpSetup, runMcpSetup} from '../mcp-server/mcp-server-setup.js';
+import {formatMcpDoctor, runMcpDoctor} from '../../features/mcp-server/mcp-server-doctor.js';
+import {formatMcpSetup, runMcpSetup} from '../../features/mcp-server/mcp-server-setup.js';
 import {readJsonBody, writeDashboardError} from './dashboard-http.js';
 import {queueDashboardTaskResponse} from './dashboard-task-commands.js';
 import {writeTaskLines} from './dashboard-task-output.js';
