+++ src/entrypoints/dashboard/dashboard-worktree-resolver.ts
