+++ src/entrypoints/dashboard/dashboard-worktree-snapshot.ts
