+++ tests/unit/dashboard-client-shell.test.ts
@@ -4,8 +4,8 @@ import {describe, expect, test} from 'vitest';
 
 describe('dashboard client shell', () => {
   test('declares the Vite entrypoint for the Preact app', () => {
-    const index = fs.readFileSync('src/features/dashboard/client/index.html', 'utf8');
-    const favicon = fs.readFileSync('src/features/dashboard/client/favicon.svg', 'utf8');
+    const index = fs.readFileSync('src/entrypoints/dashboard/client/index.html', 'utf8');
+    const favicon = fs.readFileSync('src/entrypoints/dashboard/client/favicon.svg', 'utf8');
 
     expect(index).toMatch(/<div id="dashboard-root"><\/div>/);
     expect(index).toMatch(/<script type="module" src="\.\/app\.jsx"><\/script>/);
